Actually, ListArgument should also be top-level class
